Migrating to pyramid 2.0
One of the biggest changes in Pyramid 2.0 is the change in the security
and ACL mechanism. The previous version of samba4_manager stored a
username and password for anonymous/admin login to attempt to use when
first connecting to the server, I now think that it's unnecessary and we
should be able to work with user-supplied credentials only. I also
hard-coded the version numbers of the requirements so that they are not
changed whenever this application is reinstalled.